+++ CHANGELOG.md
@@ -9,5 +9,6 @@ FEATURES:
 * **New Resource:** `panos_management_profile`
 * **New Resource:** `panos_nat_policy`
 * **New Resource:** `panos_security_policy`
+* **New Resource:** `panos_service_group`
 * **New Resource:** `panos_service_object`
 * **New Resource:** `panos_zone`